make pyproject and refactor for common methods in utils
