+++ .github/workflows/ci.yml
@@ -33,10 +33,9 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: '3.11'
-      - run: pip install pyyaml jsonschema pytest
-      - name: Set PYTHONPATH
-        run: echo "PYTHONPATH=$GITHUB_WORKSPACE" >> $GITHUB_ENV
+      - name: Install project (editable) and pytest
+        run: pip install -e . pytest
       - name: "sanity check: validate sample fixture"
-        run: python tools/validate_repo.py validate tests/fixtures/basic_repo/manifest.yml schema/v1_0_0/manifest.schema.json
+        run: python -m tools.validate_repo validate tests/fixtures/basic_repo/manifest.yml schema/v1_0_0/manifest.schema.json
       - name: Run tests
         run: pytest -q
