+++ tools/__init__.py
@@ -1,3 +1,4 @@
+from .utils import load_yaml, load_json, UniqueKeyLoader  # re-export for convenience
 # Make 'tools' a package for imports in tests/CI
